[Interop][SwiftToCxx] Moving swift::Expected to _SwiftStdlibCxxOverlay.h
