+++ lib/PrintAsClang/_SwiftStdlibCxxOverlay.h
@@ -179,4 +179,177 @@ private:
   void *_Nonnull opaqueValue = nullptr;
 };
 
+namespace _impl {
+
+constexpr static std::size_t max(std::size_t a, std::size_t b) {
+  return a > b ? a : b;
+}
+
+} // namespace _impl
+
+/// The Expected class has either an error or an value.
+template<class T>
+class Expected {
+public:
+
+  /// Default
+  constexpr Expected() noexcept {
+    new (&buffer) Error();
+    has_val = false;
+  }
+
+  constexpr Expected(const Swift::Error& error_val) noexcept {
+    new (&buffer) Error(error_val);
+    has_val = false;
+  }
+
+  constexpr Expected(const T &val) noexcept {
+    new (&buffer) T(val);
+    has_val = true;
+  }
+
+  /// Copy
+  constexpr Expected(Expected const& other) noexcept {
+    if (other.has_value())
+      new (&buffer) T(other.value());
+    else
+      new (&buffer) Error(other.error());
+
+    has_val = other.has_value();
+  }
+
+  /// Move
+  // FIXME: Implement move semantics when move Swift values is possible
+  constexpr Expected(Expected&&) noexcept { abort(); }
+
+  ~Expected() noexcept {
+    if (has_value())
+      reinterpret_cast<const T *>(buffer)->~T();
+    else
+      reinterpret_cast<Swift::Error *>(buffer)->~Error();
+  }
+
+  /// assignment
+  constexpr auto operator=(Expected&& other) noexcept = delete;
+  constexpr auto operator=(Expected&) noexcept = delete;
+
+  /// For accessing T's members
+  constexpr T const *_Nonnull operator->() const noexcept {
+    if (!has_value())
+      abort();
+    return reinterpret_cast<const T *>(buffer);
+  }
+
+  constexpr T *_Nonnull operator->() noexcept {
+    if (!has_value())
+      abort();
+    return reinterpret_cast<T *>(buffer);
+  }
+
+  /// Getting reference to T
+  constexpr T const &operator*() const & noexcept {
+    if (!has_value())
+      abort();
+    return reinterpret_cast<const T &>(buffer);
+  }
+
+  constexpr T &operator*() & noexcept {
+    if (!has_value())
+      abort();
+    return reinterpret_cast<T &>(buffer);
+  }
+
+  constexpr explicit operator bool() const noexcept { return has_value(); }
+
+  // Get value, if not exists abort
+  constexpr T const& value() const& {
+    if (!has_value())
+      abort();
+    return *reinterpret_cast<const T *>(buffer);
+  }
+
+  constexpr T& value() & {
+    if (!has_value())
+      abort();
+    return *reinterpret_cast<T *>(buffer);
+  }
+
+  // Get error
+  constexpr Swift::Error const& error() const& {
+    if (has_value())
+      abort();
+    return reinterpret_cast<const Swift::Error&>(buffer);
+  }
+
+  constexpr Swift::Error& error() & {
+    if (has_value())
+      abort();
+    return reinterpret_cast<Swift::Error&>(buffer);
+  }
+
+  constexpr bool has_value() const noexcept { return has_val; }
+
+private:
+  alignas(_impl::max(alignof(T), alignof(Swift::Error))) char buffer[_impl::max(sizeof(T), sizeof(Swift::Error))];
+  bool has_val;
+};
+
+template<>
+class Expected<void> {
+public:
+  /// Default
+  Expected() noexcept {
+    new (&buffer) Error();
+    has_val = false;
+  }
+
+  Expected(const Swift::Error& error_val) noexcept {
+    new (&buffer) Error(error_val);
+    has_val = false;
+  }
+
+
+  /// Copy
+  Expected(Expected const& other) noexcept {
+    if (other.has_value())
+      abort();
+    else
+      new (&buffer) Error(other.error());
+
+    has_val = other.has_value();
+  }
+
+  /// Move
+  // FIXME: Implement move semantics when move swift values is possible
+  [[noreturn]] Expected(Expected&&) noexcept { abort(); }
+
+  ~Expected() noexcept {
+    reinterpret_cast<Swift::Error *>(buffer)->~Error();
+  }
+
+  /// assignment
+  constexpr auto operator=(Expected&& other) noexcept = delete;
+  constexpr auto operator=(Expected&) noexcept = delete;
+
+
+  constexpr explicit operator bool() const noexcept { return has_value(); }
+
+  // Get error
+  constexpr Swift::Error const& error() const& {
+    if (has_value())
+      abort();
+    return reinterpret_cast<const Swift::Error&>(buffer);
+  }
+
+  constexpr Swift::Error& error() & {
+    if (has_value())
+      abort();
+    return reinterpret_cast<Swift::Error&>(buffer);
+  }
+
+  constexpr bool has_value() const noexcept { return has_val; }
+private:
+  alignas(alignof(Swift::Error)) char buffer[sizeof(Swift::Error)];
+  bool has_val;
+};
 #endif
