+++ test/Interop/SwiftToCxx/functions/swift-expected-execution.cpp
@@ -1,8 +1,8 @@
 // RUN: %empty-directory(%t)
 
-// RUN: %target-swift-frontend %S/swift-functions-errors.swift -typecheck -module-name Functions -clang-header-expose-decls=has-expose-attr -emit-clang-header-path %t/functions.h
+// RUN: %target-swift-frontend %S/swift-functions-errors.swift -typecheck -module-name Functions -Xcc -fno-exceptions -enable-experimental-cxx-interop -emit-clang-header-path %t/functions.h
 
-// RUN: %target-interop-build-clangxx -c %s -I %t -o %t/swift-expected-execution.o
+// RUN: %target-interop-build-clangxx -fno-exceptions -c %s -I %t -fno-exceptions -o %t/swift-expected-execution.o
 // RUN: %target-interop-build-swift %S/swift-functions-errors.swift -o %t/swift-expected-execution -Xlinker %t/swift-expected-execution.o -module-name Functions -Xfrontend -entry-point-function-name -Xfrontend swiftMain
 
 // RUN: %target-codesign %t/swift-expected-execution
@@ -17,14 +17,14 @@
 int main() {
 
   // Test Empty Constructor
-  auto testIntEmpty = swift::Expected<int>();
+  auto testIntEmpty = Swift::Expected<int>();
 
   // Test Error Constructor
-  swift::Error e;
-  auto testIntError = swift::Expected<int>(e);
+  Swift::Error e;
+  auto testIntError = Swift::Expected<int>(e);
 
   // Test Value Constructor
-  auto testIntValue = swift::Expected<int>(42);
+  auto testIntValue = Swift::Expected<int>(42);
 
   // Test Copy Constructor
   auto testCopy = testIntEmpty;
@@ -32,9 +32,9 @@ int main() {
   // TODO: Test Move Constructor
 
   // Test Destructor
-  auto testDestEmpty = swift::Expected<int>();
-  auto testDestInt = swift::Expected<int>(42);
-  auto testDestError = swift::Expected<int>(e);
+  auto testDestEmpty = Swift::Expected<int>();
+  auto testDestInt = Swift::Expected<int>(42);
+  auto testDestError = Swift::Expected<int>(e);
   testDestEmpty.~Expected();
   testDestInt.~Expected();
   testDestError.~Expected();
@@ -73,23 +73,22 @@ int main() {
   if (testIntValue.value() == 42)
     printf("Test get T's Value\n");
 
-  // Test get swift::Error (const)
-  try {
-    Functions::throwFunction();
-  } catch (swift::Error& e) {
-    const auto errorExp = swift::Expected<int>(e);
-    auto err = errorExp.error(); // using const function
-    auto errorVal = err.as<Functions::NaiveErrors>();
-    errorVal.getMessage();
+  // Test get Swift::Error (const)
+  const auto constExpectedResult = Functions::throwFunctionWithPossibleReturn(0);
+  if (!constExpectedResult.has_value()) {
+    auto constError = constExpectedResult.error(); // using const function
+    auto optionalError = constError.as<Functions::NaiveErrors>();
+    auto errorValue = optionalError.get();
+    errorValue.getMessage();
   }
 
-  // Test get swift::Error
-  try {
-    Functions::throwFunction();
-  } catch (swift::Error& e) {
-    auto err = swift::Expected<int>(e).error();
-    auto errorVal = err.as<Functions::NaiveErrors>();
-    errorVal.getMessage();
+  // Test get Swift::Error
+  auto expectedResult = Functions::throwFunctionWithPossibleReturn(0);
+  if (!expectedResult.has_value()) {
+    auto error = expectedResult.error();
+    auto optionalError = error.as<Functions::NaiveErrors>();
+    auto errorValue = optionalError.get();
+    errorValue.getMessage();
   }
 
   // Test has Value
@@ -108,9 +107,9 @@ int main() {
 // CHECK-NEXT: Test operator bool
 // CHECK-NEXT: Test get T's Value (const)
 // CHECK-NEXT: Test get T's Value
-// CHECK-NEXT: passThrowFunction
-// CHECK-NEXT: throwError
-// CHECK-NEXT: passThrowFunction
-// CHECK-NEXT: throwError
+// CHECK-NEXT: passThrowFunctionWithPossibleReturn
+// CHECK-NEXT: returnError
+// CHECK-NEXT: passThrowFunctionWithPossibleReturn
+// CHECK-NEXT: returnError
 // CHECK-NEXT: testIntValue has a value
 // CHECK-NEXT: testIntError doesn't have a value
